refactor: change processor module structure
