+++ keriox_core/src/component/mod.rs
@@ -1,259 +1,27 @@
-use std::{convert::TryFrom, path::Path, sync::Arc};
+use std::{path::Path, sync::Arc};
 
 use crate::{
     database::sled::SledEventDatabase,
-    derivation::{basic::Basic, self_addressing::SelfAddressing, self_signing::SelfSigning},
     error::Error,
-    event::receipt::Receipt,
-    event_message::{
-        event_msg_builder::ReceiptBuilder,
-        key_event_message::KeyEvent,
-        serialization_info::SerializationFormats,
-        signed_event_message::{Message, SignedNontransferableReceipt},
-        EventMessage,
-    },
-    event_parsing::message::signed_event_stream,
-    oobi::{LocationScheme, OobiManager, Role},
-    prefix::{BasicPrefix, IdentifierPrefix},
-    processor::{
-        event_processor::Processor, event_storage::EventStorage, notification::Notification,
-        responder::Responder,
-    },
+    event_message::{serialization_info::SerializationFormats, signed_event_message::Message},
+    prefix::IdentifierPrefix,
+    processor::{event_storage::EventStorage, Processor},
     query::{
         key_state_notice::KeyStateNotice,
         query_event::{QueryData, SignedQuery},
-        reply_event::{ReplyEvent, ReplyRoute, SignedReply},
         ReplyType,
     },
-    signer::Signer,
     state::IdentifierState,
 };
 
-pub struct NontransferableActor<P: Processor> {
-    pub prefix: BasicPrefix,
-    pub actor: Actor<P>,
-    pub signer: Arc<Signer>,
-    pub oobi_manager: Arc<OobiManager>,
-    responder: Arc<Responder<Notification>>,
-}
+pub mod nontransferable_component;
 
-impl<P: Processor> NontransferableActor<P> {
-    pub fn setup(
-        public_address: url::Url,
-        event_db_path: &Path,
-        oobi_db_path: &Path,
-        priv_key: Option<String>,
-    ) -> Result<Self, Error> {
-        let oobi_manager = Arc::new(OobiManager::new(oobi_db_path));
-        let signer = priv_key
-            .map(|key| Signer::new_with_seed(&key.parse()?))
-            .unwrap_or(Ok(Signer::new()))?;
-        let prefix = Basic::Ed25519.derive(signer.public_key());
-        let witness = Actor::<P>::new(event_db_path)?;
-        // construct witness loc scheme oobi
-        let loc_scheme = LocationScheme::new(
-            IdentifierPrefix::Basic(prefix.clone()),
-            public_address.scheme().parse().unwrap(),
-            public_address.clone(),
-        );
-        let reply = ReplyEvent::new_reply(
-            ReplyRoute::LocScheme(loc_scheme),
-            SelfAddressing::Blake3_256,
-            SerializationFormats::JSON,
-        )?;
-        let signed_reply = SignedReply::new_nontrans(
-            reply.clone(),
-            prefix.clone(),
-            SelfSigning::Ed25519Sha512.derive(signer.sign(reply.serialize()?)?),
-        );
-        oobi_manager.save_oobi(signed_reply)?;
-        let responder = Arc::new(Responder::new());
-        Ok(Self {
-            prefix,
-            actor: witness,
-            oobi_manager,
-            signer: Arc::new(signer),
-            responder,
-        })
-    }
-
-    pub fn respond(&self, signer: Arc<Signer>) -> Result<Vec<Message>, Error> {
-        let response = Vec::new();
-        while let Some(event) = self.responder.get_data_to_respond() {
-            match event {
-                Notification::KeyEventAdded(event) => {
-                    let non_trans_receipt =
-                        self.respond_to_key_event(event.event_message, signer.clone())?;
-                    self.actor.storage.add_mailbox_receipt(non_trans_receipt)?;
-                }
-                _ => return Err(Error::SemanticError("Wrong notification type".into())),
-            }
-        }
-        Ok(response)
-    }
-
-    fn respond_to_key_event(
-        &self,
-        event_message: EventMessage<KeyEvent>,
-        signer: Arc<Signer>,
-    ) -> Result<SignedNontransferableReceipt, Error> {
-        // Create witness receipt and add it to db
-        let ser = event_message.serialize()?;
-        let signature = signer.sign(&ser)?;
-        let rcp = ReceiptBuilder::default()
-            .with_receipted_event(event_message)
-            .build()?;
-
-        let signature = SelfSigning::Ed25519Sha512.derive(signature);
-
-        let signed_rcp = SignedNontransferableReceipt::new(
-            &rcp,
-            Some(vec![(self.prefix.clone(), signature)]),
-            None,
-        );
-
-        self.actor
-            .processor
-            .process(Message::NontransferableRct(signed_rcp.clone()))?;
-        Ok(signed_rcp)
-    }
-
-    pub fn get_signed_ksn_for_prefix(
-        &self,
-        prefix: &IdentifierPrefix,
-        signer: Arc<Signer>,
-    ) -> Result<SignedReply, Error> {
-        let ksn = self.actor.get_ksn_for_prefix(prefix)?;
-        let rpy = ReplyEvent::new_reply(
-            ReplyRoute::Ksn(IdentifierPrefix::Basic(self.prefix.clone()), ksn),
-            SelfAddressing::Blake3_256,
-            SerializationFormats::JSON,
-        )?;
-
-        let signature = SelfSigning::Ed25519Sha512.derive(signer.sign(&rpy.serialize()?)?);
-        Ok(SignedReply::new_nontrans(
-            rpy,
-            self.prefix.clone(),
-            signature,
-        ))
-    }
-
-    pub fn get_end_role_for_id(
-        &self,
-        cid: &IdentifierPrefix,
-        role: Role,
-    ) -> Result<Option<Vec<SignedReply>>, Error> {
-        Ok(self.oobi_manager.get_end_role(cid, role).unwrap())
-    }
-
-    pub fn get_loc_scheme_for_id(
-        &self,
-        eid: &IdentifierPrefix,
-    ) -> Result<Option<Vec<SignedReply>>, Error> {
-        Ok(match self.oobi_manager.get_loc_scheme(eid).unwrap() {
-            Some(oobis_to_sign) => Some(
-                oobis_to_sign
-                    .iter()
-                    .map(|oobi_to_sing| {
-                        let signature =
-                            self.signer.sign(oobi_to_sing.serialize().unwrap()).unwrap();
-                        SignedReply::new_nontrans(
-                            oobi_to_sing.clone(),
-                            self.prefix.clone(),
-                            SelfSigning::Ed25519Sha512.derive(signature),
-                        )
-                    })
-                    .collect(),
-            ),
-            None => None,
-        })
-    }
-
-    // Returns messages if they can be returned immediately, i.e. for query message
-    pub fn process(&self, msg: Message) -> Result<Vec<Message>, Error> {
-        let mut responses = Vec::new();
-        match msg.clone() {
-            Message::Event(ev) => {
-                self.actor.process(&msg).unwrap();
-                // check if receipts are attached
-                if let Some(witness_receipts) = ev.witness_receipts {
-                    // Create and process witness receipts
-                    // TODO What timestamp should be set?
-                    let id = ev.event_message.event.get_prefix();
-                    let receipt = Receipt {
-                        receipted_event_digest: ev.event_message.get_digest(),
-                        prefix: id,
-                        sn: ev.event_message.event.get_sn(),
-                    };
-                    let signed_receipt = SignedNontransferableReceipt::new(
-                        &receipt.to_message(SerializationFormats::JSON).unwrap(),
-                        None,
-                        Some(witness_receipts),
-                    );
-                    self.actor
-                        .process(&Message::NontransferableRct(signed_receipt))
-                        .unwrap();
-                }
-            }
-            Message::Query(qry) => {
-                let response = self.actor.process_signed_query(qry).unwrap();
-                match response {
-                    ReplyType::Ksn(ksn) => {
-                        let rpy = ReplyEvent::new_reply(
-                            ReplyRoute::Ksn(IdentifierPrefix::Basic(self.prefix.clone()), ksn),
-                            SelfAddressing::Blake3_256,
-                            SerializationFormats::JSON,
-                        )?;
-
-                        let signature =
-                            SelfSigning::Ed25519Sha512.derive(self.signer.sign(&rpy.serialize()?)?);
-                        let reply = Message::Reply(SignedReply::new_nontrans(
-                            rpy,
-                            self.prefix.clone(),
-                            signature,
-                        ));
-                        responses.push(reply);
-                    }
-                    ReplyType::Kel(msgs) | ReplyType::Mbx(msgs) => responses.extend(msgs),
-                };
-            }
-            Message::NontransferableRct(_) => self.actor.process(&msg).unwrap(),
-            Message::TransferableRct(_) => self.actor.process(&msg).unwrap(),
-            Message::Reply(reply) => match reply.reply.get_route() {
-                ReplyRoute::Ksn(_, _) => self.actor.process(&msg).unwrap(),
-                ReplyRoute::LocScheme(_)
-                | ReplyRoute::EndRoleAdd(_)
-                | ReplyRoute::EndRoleCut(_) => self.oobi_manager.process_oobi(reply).unwrap(),
-            },
-        };
-        Ok(responses)
-    }
-
-    pub fn parse_and_process(&self, input_stream: &[u8]) -> Result<Vec<Message>, Error> {
-        let (_, msgs) = signed_event_stream(input_stream)
-            .map_err(|e| Error::DeserializeError(e.to_string()))
-            .unwrap();
-
-        let output = msgs
-            .into_iter()
-            .map(|msg| -> Result<_, _> {
-                let msg = Message::try_from(msg)?;
-                self.process(msg)
-            })
-            // TODO: avoid unwrap
-            .map(|d| d.unwrap())
-            .flatten()
-            .collect();
-        Ok(output)
-    }
-}
-
-pub struct Actor<P: Processor> {
+pub struct Component<P: Processor> {
     processor: P,
     pub storage: EventStorage,
 }
 
-impl<P: Processor> Actor<P> {
+impl<P: Processor> Component<P> {
     pub fn new(path: &Path) -> Result<Self, Error> {
         let (processor, storage) = {
             let witness_db = Arc::new(SledEventDatabase::new(path)?);
@@ -268,10 +36,6 @@ impl<P: Processor> Actor<P> {
 
     pub fn get_db_ref(&self) -> Arc<SledEventDatabase> {
         self.storage.db.clone()
-    }
-
-    pub fn process(&self, msg: &Message) -> Result<(), Error> {
-        self.processor.process(msg.clone())
     }
 
     pub fn get_kel_for_prefix(&self, id: &IdentifierPrefix) -> Result<Option<Vec<u8>>, Error> {
@@ -295,6 +59,11 @@ impl<P: Processor> Actor<P> {
             .ok_or_else(|| Error::SemanticError("No state in db".into()))?;
         let ksn = KeyStateNotice::new_ksn(state, SerializationFormats::JSON);
         Ok(ksn)
+    }
+
+    /// Process for events that updates database
+    pub fn process(&self, msg: &Message) -> Result<(), Error> {
+        self.processor.process(msg.clone())
     }
 
     pub fn process_signed_query(&self, qr: SignedQuery) -> Result<ReplyType, Error> {
@@ -344,6 +113,11 @@ impl<P: Processor> Actor<P> {
 
 #[test]
 pub fn test_ksn_query() -> Result<(), Error> {
+    use crate::{
+        component::nontransferable_component::NontransferableComponent,
+        event_message::signed_event_message::Message, prefix::IdentifierPrefix,
+        query::reply_event::ReplyRoute,
+    };
     use std::convert::TryFrom;
 
     use crate::event_parsing::message::signed_message;
@@ -354,7 +128,7 @@ pub fn test_ksn_query() -> Result<(), Error> {
     let root = Builder::new().prefix("test-db").tempdir().unwrap();
     let oobi_root = Builder::new().prefix("test-db2").tempdir().unwrap();
 
-    let witness = NontransferableActor::<WitnessProcessor>::setup(
+    let witness = NontransferableComponent::<WitnessProcessor>::setup(
         url::Url::parse("http://localhost").unwrap(),
         root.path(),
         oobi_root.path(),
