+++ keriox_core/src/processor/basic_processor.rs
@@ -1,17 +1,18 @@
 use std::sync::Arc;
 
-use crate::{
-    database::sled::SledEventDatabase,
-    error::Error,
-    event_message::signed_event_message::{Message, SignedEventMessage},
-    processor::notification::{JustNotification, NotificationBus},
+use crate::event_message::signed_event_message::{Message, SignedEventMessage};
+#[cfg(feature = "query")]
+use crate::{database::sled::SledEventDatabase, error::Error};
+
+use super::{
+    notification::{Notification, NotificationBus},
+    validator::EventValidator,
+    EventProcessor, Processor,
 };
 
-use super::{notification::Notification, validator::EventValidator, EventProcessor, Processor};
+pub struct BasicProcessor(EventProcessor);
 
-pub struct WitnessProcessor(EventProcessor);
-
-impl Processor for WitnessProcessor {
+impl Processor for BasicProcessor {
     fn process(&self, message: Message) -> Result<(), Error> {
         self.process(message)
     }
@@ -21,26 +22,13 @@ impl Processor for WitnessProcessor {
     }
 }
 
-impl WitnessProcessor {
+impl BasicProcessor {
     pub fn new(db: Arc<SledEventDatabase>) -> Self {
-        use crate::processor::escrow::{OutOfOrderEscrow, PartiallySignedEscrow};
-        let mut bus = NotificationBus::new();
-        bus.register_observer(
-            Arc::new(PartiallySignedEscrow::new(db.clone())),
-            vec![JustNotification::PartiallySigned],
-        );
-        bus.register_observer(
-            Arc::new(OutOfOrderEscrow::new(db.clone())),
-            vec![
-                JustNotification::OutOfOrder,
-                JustNotification::KeyEventAdded,
-            ],
-        );
-        let processor = EventProcessor::new(db, bus);
+        let processor = EventProcessor::new(db, NotificationBus::default());
         Self(processor)
     }
 
-    fn witness_processing_strategy(
+    fn basic_processing_strategy(
         db: Arc<SledEventDatabase>,
         publisher: &NotificationBus,
         signed_event: SignedEventMessage,
@@ -56,8 +44,7 @@ impl WitnessProcessor {
                 publisher.notify(&Notification::OutOfOrder(signed_event))
             }
             Err(Error::NotEnoughReceiptsError) => {
-                db.add_kel_finalized_event(signed_event.clone(), id)?;
-                publisher.notify(&Notification::KeyEventAdded(signed_event))
+                publisher.notify(&Notification::PartiallyWitnessed(signed_event))
             }
             Err(Error::NotEnoughSigsError) => {
                 publisher.notify(&Notification::PartiallySigned(signed_event))
@@ -76,7 +63,7 @@ impl WitnessProcessor {
     /// Ignore not fully witness error and accept not fully witnessed events.
     pub fn process(&self, message: Message) -> Result<(), Error> {
         self.0
-            .process(message, WitnessProcessor::witness_processing_strategy)?;
+            .process(message, BasicProcessor::basic_processing_strategy)?;
         Ok(())
     }
 }
