+++ keriox_core/src/processor/mod.rs
@@ -1,17 +1,9 @@
 use std::sync::Arc;
 
-use crate::event_message::signed_event_message::{Message, SignedEventMessage};
-#[cfg(feature = "query")]
-use crate::{
-    database::sled::SledEventDatabase, error::Error,
-    event_message::signed_event_message::TimestampedSignedEventMessage, prefix::IdentifierPrefix,
-    state::IdentifierState,
-};
-
 // #[cfg(feature = "async")]
 // pub mod async_processing;
+pub mod basic_processor;
 pub mod escrow;
-pub mod event_processor;
 pub mod event_storage;
 pub mod notification;
 pub mod responder;
@@ -20,67 +12,113 @@ mod tests;
 pub mod validator;
 pub mod witness_processor;
 
+use crate::{
+    database::sled::SledEventDatabase,
+    error::Error,
+    event_message::signed_event_message::{
+        Message, SignedEventMessage, TimestampedSignedEventMessage,
+    },
+    prefix::IdentifierPrefix,
+    query::reply_event::ReplyRoute,
+    state::IdentifierState,
+};
+
 use self::{
-    event_processor::{EventProcessor, Processor},
     notification::{Notification, NotificationBus},
     validator::EventValidator,
 };
 
-pub struct BasicProcessor(EventProcessor);
-
-impl Processor for BasicProcessor {
-    fn process(&self, message: Message) -> Result<(), Error> {
-        self.process(message)
-    }
-
-    fn new(db: Arc<SledEventDatabase>) -> Self {
-        Self::new(db)
-    }
+pub trait Processor {
+    fn new(db_path: Arc<SledEventDatabase>) -> Self;
+    fn process(&self, message: Message) -> Result<(), Error>;
 }
 
-impl BasicProcessor {
-    pub fn new(db: Arc<SledEventDatabase>) -> Self {
-        let processor = EventProcessor::new(db, NotificationBus::default());
-        Self(processor)
-    }
+pub struct EventProcessor {
+    db: Arc<SledEventDatabase>,
+    validator: EventValidator,
+    publisher: NotificationBus,
+}
 
-    fn basic_processing_strategy(
-        db: Arc<SledEventDatabase>,
-        publisher: &NotificationBus,
-        signed_event: SignedEventMessage,
-    ) -> Result<(), Error> {
-        let id = &signed_event.event_message.event.get_prefix();
+impl EventProcessor {
+    pub fn new(db: Arc<SledEventDatabase>, publisher: NotificationBus) -> Self {
         let validator = EventValidator::new(db.clone());
-        match validator.validate_event(&signed_event) {
-            Ok(_) => {
-                db.add_kel_finalized_event(signed_event.clone(), id)?;
-                publisher.notify(&Notification::KeyEventAdded(signed_event))
-            }
-            Err(Error::EventOutOfOrderError) => {
-                publisher.notify(&Notification::OutOfOrder(signed_event))
-            }
-            Err(Error::NotEnoughReceiptsError) => {
-                publisher.notify(&Notification::PartiallyWitnessed(signed_event))
-            }
-            Err(Error::NotEnoughSigsError) => {
-                publisher.notify(&Notification::PartiallySigned(signed_event))
-            }
-            Err(Error::EventDuplicateError) => {
-                db.add_duplicious_event(signed_event.clone(), id)?;
-                publisher.notify(&Notification::DupliciousEvent(signed_event))
-            }
-            Err(e) => Err(e),
+        Self {
+            db,
+            validator,
+            publisher,
         }
     }
 
     /// Process
     ///
-    /// Process a deserialized KERI message.
-    /// Ignore not fully witness error and accept not fully witnessed events.
-    pub fn process(&self, message: Message) -> Result<(), Error> {
-        self.0
-            .process(message, BasicProcessor::basic_processing_strategy)?;
-        Ok(())
+    /// Process a deserialized KERI message
+    /// Update database based on event validation result.
+    pub fn process<F>(&self, message: Message, processing_strategy: F) -> Result<(), Error>
+    where
+        F: Fn(Arc<SledEventDatabase>, &NotificationBus, SignedEventMessage) -> Result<(), Error>,
+    {
+        match message {
+            Message::Event(signed_event) => {
+                processing_strategy(self.db.clone(), &self.publisher, signed_event)
+            }
+            Message::NontransferableRct(rct) => {
+                let id = &rct.body.event.prefix;
+                match self.validator.validate_witness_receipt(&rct) {
+                    Ok(_) => {
+                        self.db.add_receipt_nt(rct.to_owned(), id)?;
+                        self.publisher.notify(&Notification::ReceiptAccepted)
+                    }
+                    Err(Error::MissingEvent) => self
+                        .publisher
+                        .notify(&Notification::ReceiptOutOfOrder(rct.clone())),
+                    Err(e) => Err(e),
+                }
+            }
+            Message::TransferableRct(vrc) => {
+                match self.validator.validate_validator_receipt(&vrc) {
+                    Ok(_) => {
+                        self.db.add_receipt_t(vrc.clone(), &vrc.body.event.prefix)?;
+                        self.publisher.notify(&Notification::ReceiptAccepted)
+                    }
+                    Err(Error::MissingEvent) | Err(Error::EventOutOfOrderError) => self
+                        .publisher
+                        .notify(&Notification::TransReceiptOutOfOrder(vrc.clone())),
+                    Err(e) => Err(e),
+                }
+            }
+            #[cfg(feature = "query")]
+            Message::Reply(rpy) => match rpy.reply.get_route() {
+                ReplyRoute::Ksn(_, _) => match self.validator.process_signed_ksn_reply(&rpy) {
+                    Ok(_) => {
+                        self.db
+                            .update_accepted_reply(rpy.clone(), &rpy.reply.get_prefix())
+                        // self.publisher.notify(&Notification::ReplyUpdated)
+                    }
+                    Err(Error::EventOutOfOrderError) => {
+                        self.publisher.notify(&Notification::KsnOutOfOrder(rpy))
+                    }
+                    Err(anything) => Err(anything),
+                },
+                #[cfg(feature = "oobi")]
+                ReplyRoute::EndRoleAdd(_)
+                | ReplyRoute::EndRoleCut(_)
+                | ReplyRoute::LocScheme(_) => {
+                    // check signature
+                    self.validator
+                        .verify(&rpy.reply.serialize()?, &rpy.signature)?;
+                    // check digest
+                    rpy.reply.check_digest()?;
+                    // self.publisher.notify(&Notification::GotOobi(rpy))
+                    Ok(())
+                }
+            },
+            #[cfg(feature = "query")]
+            Message::Query(_) => {
+                // TODO should do nothing?
+                // It doesn't update database
+                Ok(())
+            }
+        }
     }
 }
